Added files via upload
index file and css file as they are at the moment.
There is stille something wich is not working, due to the :hover state and :focus state addede, but not working on touch devices.
I am going to figure that out shortly. It is hard, when I can't use jQuery og just JavaScript.